keyboard_scan: slow down polling rate when not debouncing
The keyboard scanning task polls continuously as long as any button is
pressed. This is necessary as the idling detection requires all butons
to be released, however, we only need dense data points to debounce
the matrix (or rather, deglitching it).

Implement a new setting for slowing down the polling rate when the
matrix is stable, which should help reducing the keyboard task CPU
utilization when buttons are pressed but not changing.

Set the default tune it to 9ms stable scans (normal is 3ms), update all
existing boards to use the new timings if the previous setting was the
default, maintain the existing behavior if the settig was tuned (many
defaults to 10ms already, some to 1.5ms).

The number is somewhat arbitrary at this point but it seems like a
reasonable starting point.

The new field may end up being unset on overridden config that do not
have the new value, in that case the scan would default to use the
debounce delay all the time (i.e. the current behavior).

A col2 waveform with this looks something like
https://screenshot.googleplex.com/3c7ZqSNqRygXB7B

BUG=b:376670220
TEST=build and run on brex, captured some waveform, typed a bunch of
stuff, checked all keys with ksstate as well

Change-Id: Ia89c75a02febfcba994b091560fbe1ffde0a5ee2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/5966734
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Yuval Peress <peress@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6712310
Commit-Queue: Rob Barnes <robbarnes@google.com>
Reviewed-by: Boris Mittelberg <bmbm@google.com>
Tested-by: Rob Barnes <robbarnes@google.com>