common: motion_sense: Mark sensor as NOT_INITIALIZED as soon as possible
When a sensor is about to lose power, mark its state as NOT_INITIALIZED.
It does not prevent current task to run, and we are still scheduling
the `switch_sensor_rate` in the hook task.
But we will ignore further call to process sensor data and we are
guarantee to re-initialize the sensor if suspend and resume events
happen in short order.

BUG=b:377845255
TEST=To be tested.

Change-Id: I791d09129cebefa0869bfe5d51ab04092e1c8ab1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6001370
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/6517733